feat(stress-test): scenario 1 exécuté — 4 pipelines × 16 appels LLM
- metrics_report.json : 10 métriques pour OTel / Data Mesh / D-SIG / OTel→D-SIG
- llm_responses.json : 16 évaluations LLM (claude-sonnet-4-6) avec input_tokens_to_llm
- summary.csv : tableau plat prêt pour DeepSeek
- PROGRESSION.md : état du projet et résultats bruts
- Fix : process_count float pour accepter NaN (silences)
- Fix : loader .env avec affectation directe (surcharge nix-shell)
- Ajout : .gitignore, .env.example, shell.nix

+++ dsig-stress-test/scenario1/data/fetch_data.py
@@ -121,7 +121,7 @@ def generate_synthetic(gt):
         mem    = _baseline_series(n, 40, 70, seed_base + 2)
         lat    = _baseline_series(n, 3,  8,  seed_base + 3)
         disk   = _baseline_series(n, 5,  50, seed_base + 4)
-        procs  = _baseline_series(n, 80, 200, seed_base + 5).astype(int)
+        procs  = _baseline_series(n, 80, 200, seed_base + 5).astype(float)  # float pour accepter NaN
         uptime = np.arange(1, n + 1, dtype=float) * 60  # seconds since boot
 
         # -- Inject incidents per ground_truth --
